backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
xgboost
scikit-learn
numpy
snowflake-connector-python
